Improve of whitelisting in privileged tests
This change updates privileged test to consider
only containers from CNF under test. As a result
of this change, the 'allowlist_helm_chart_container_names'
parameter in the cnf-testsuite.yml configuration file
of lot of examples and samples is no needed. Therefore
this patch also removes this parameter from cnf-testsuite.yml
in examples and samples where it is no need anymore.

Closes-Bug: #1906
Signed-off-by: horecoli <oliver.horecny@tietoevry.com>