Prevent microservice tests from passing on errors
Refs: #1959
This updates the behaviour so the test will not pass
when there is an environment or external error causing
the test to not be able to do a complete validation
as intended. (eg. the containers could not be inspected
because of an external unrelated issue).
The error is reported and the test will not pass.

Signed-off-by: Martin Matyas <martin.matyas@tietoevry.com>