workflow: Cancel workflows on push
In our CICD, when new commit is pushed (or amended) while
workflow for old one is still running - there is no auto-cancelation.
Old workflows have to be canceled manually by maintainers
to avoid unneccessary waiting.

This change fixes that issue by grouping workflows by PR number or
branch ref (in case for push event), and if new workflow is ran in the
same group - old one would be automatically cancelled.

Signed-off-by: Konstantin Yarovoy <konstantin.yarovoy@tietoevry.com>